First pass: Remove all actions from ensemble and single
