+++ callflow/algorithms/__init__.py
@@ -1,1 +1,3 @@
 from .k_medoids import KMedoids
+from .deltacon_similarity import DeltaConSimilarity
+
