+++ callflow/ensemble/__init__.py
@@ -1,4 +1,3 @@
-from .actions import *
 from .cct import EnsembleCCT
 from .supergraph import EnsembleSuperGraph
 from .callflow import EnsembleCallFlow
