--- callflow/ensemble/actions/__init__.py
@@ -1,10 +1,0 @@
-from .module_hierarchy import ModuleHierarchy
-from .parameter_projection import ParameterProjection
-from .scatterplot import DiffScatterplot
-from .similarity import Similarity
-from .union_graph import UnionGraph
-from .gradients import KDE_gradients
-from .auxiliary import Auxiliary
-from .compare import Compare
-from .split_callee import SplitCallee
-from .split_caller import SplitCaller
