+++ callflow/modules/__init__.py
@@ -1,4 +1,14 @@
-from .auxiliary import Auxiliary
+from .auxiliary_ensemble import EnsembleAuxiliary
+from .auxiliary_single import SingleAuxiliary
+from .auxiliary_hd5 import FastEnsembleAuxiliary
 from .boxplot import BoxPlot
 from .rank_histogram import RankHistogram
+from .mini_histogram import MiniHistogram
+from .tooltip import ToolTip
+from .function_list import FunctionList
 from .generic_histogram import GenericHistogram
+from .gradients import Gradients
+from .runtime_scatterplot import RuntimeScatterplot
+from .parameter_projection import ParameterProjection
+from .compare import Compare
+from .module_hierarchy import ModuleHierarchy
