+++ callflow/single/__init__.py
@@ -1,4 +1,3 @@
-from .actions import *
 from .cct import SingleCCT
 from .supergraph import SingleSuperGraph
 from .callflow import SingleCallFlow
