--- callflow/single/actions/__init__.py
@@ -1,4 +1,0 @@
-from .auxiliary import *
-from .mini_histogram import *
-from .scatterplot import *
-from .function_list import *
